Update syncState to support multiple keys at a time
Previously, when setState was replaced it overwrote all the previously bound "setState"s so only the last syncState actually wrote to Firebase. My change chains all the setState calls together and fixes this. I am a little unclear on what the `sync.called` if statement was intended to achieve, but removing it causes the library to function as expected (I may have missed an edge case though).